requirements.txt
PyQt5
PyOpenGL
Pillow
numpy